{"schema":"decodex/commit/1","summary":"split rust utility modules","authority":"manual"}
